requirements.txt
django >= 4.2
django-csp >= 3.6
django-recaptcha @ git+https://github.com/alanverresen/django-recaptcha.git@issue101-externalscripts
